requirements.txt
azure-storage-file-datalake>=12.0.0
azure-identity>=1.5.0
apache-airflow>=2.0.0
